pythos/augur/pyproject.toml
[project]
name = "augur"
version = "0.1.0"
description = "AUGUR: Airflow United Geospatial Universal Reader"
authors = [
    {name = "eric rodefeld",email = "eric.d.rodefeld@gmail.com"},
]
dependencies = [
    "django==5.1",
    "factory-boy>=3.3.1",
    "django-ninja>=1.3.0",
    "django-extensions>=3.2.3",
    "requests>=2.32.3",
    "psycopg[binary,pool]>=3.2.3",
    "psycopg-binary>=3.2.3",
    "geojson-pydantic>=1.1.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}

[tool.setuptools]
py-modules = []
